Introduce solvers
Register a solver per puzzle in the container. The solvers can make use of concrete classes to solve the puzzles.
The concrete classes are reuseable. They could be injected in the solvers as well, but since, at this point, it seems the core logic doesn't depend on external sources, and it doesn't make much sense to test parts of a puzzle without those concrete classes, I didn't go that way.
But the solvers are registered, and this allows me to test the solver for the specific input format of the puzzle as well as the test values, and it allows me to automatically run solvers based on a folder containing files with the actual input of each puzzle.

Yes, at this point this seems like over-engineering. But this investment will pay itself back within 200 years of doing AoC, mark my words!

Btw, also contains the implementaion of puzzles 1 and 2 of Day 1.