Refactoring: Prepare for suspendable programs.
We need to run multiple programs that can provide feedback to each other. This could be done using threads, but instead I'm gonna try to allow programs to be suspended, so they can run turn based on a single thread in the puzzle solver, and on a single virtual IntCode processor, even.

So, instead of just executing code, keep the program state (code/memory, instruction pointer and state) in a separate class.
The processor can Run a program until it's halted or suspended.
The existing execute method constructs a local program, and calls the new Run method in a loop until it the program is halted. This makes it behave the same as it did before.
The intcode processor cannot suspend a program just yet. It will just read from the IO synchronously as it did before, but with this change, the infrastructure is there.