Suspend programs that are waiting for input
The state of the program is now completely in TIntCodeProgram. The processor just executes (actually, I  wonder whether that logic could be in the same class as TProgram, but well...)

When instruction 3 (read input) is encountered, the program goes to suspended state. It's up to the intiator of the program to prepare some input then, and resume the program.

Consideration: At this moment, the program just suspends. There is no actual event that it waits for.This could be implemented in a more advanced way by having some kind of input device that the program can wait for. The device (attached to the processor), can then be signaled by the program, the program will suspend after that, and the device can signal the processor to resume running the program once it has some input ready. We'll see if we need that later. ;)