Optimize day 12, 2, finding the Least Common Multiple
This is still a brute force version that theoretically compares a list of multiples to see what is the lowest multiple that is shared by each.
But instead of building an actual list, a repetition of while loops just adds multiplies (by adding) each number that is smaller than the next, until they are all equal. The overhead of such while loops is minimal. The remaining slowness is partially because of Int64 comparisons on a 32 bit compiled program (looks like that's doubling the number of instructions), + the sheer number of iterations.

My answer was about 307 quadrillion. In this code, the repeat loop interated about 1 billion times, each time checking whether to enter each of the 3 while loops. In total they were entered 1.6 billion times, doing that many additions.

This all ran in about 15 seconds, where the previous version took close to 10 minutes.